Finalize version 1 cleanup

+++ Components/Pages/TodosPage.razor.cs
@@ -1,5 +1,6 @@
 using AzubiLog.Services.Todos;
 using Microsoft.AspNetCore.Components;
+using Microsoft.JSInterop;
 
 namespace AzubiLog.Components.Pages;
 
@@ -7,6 +8,9 @@ public partial class TodosPage : ComponentBase
 {
     [Inject]
     private ITodoService TodoService { get; set; } = null!;
+
+    [Inject]
+    private IJSRuntime JsRuntime { get; set; } = null!;
 
     protected TodoPageViewModel? ViewModel { get; private set; }
 
@@ -34,6 +38,14 @@ public partial class TodosPage : ComponentBase
 
     protected async Task HandleDeleteAsync(int todoId)
     {
+        var confirmed = await JsRuntime.InvokeAsync<bool>(
+            "confirm",
+            Localizer["TodoDeleteConfirm"].Value);
+        if (!confirmed)
+        {
+            return;
+        }
+
         await TodoService.DeleteTodoAsync(todoId);
         await ReloadAsync();
     }
